requirements.txt
absl-py==1.0.0
apache-beam==2.37.0
appdirs==1.4.4
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==21.4.0
audioread==2.1.9
avro-python3==1.10.2
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
contextlib2==21.6.0
crcmod==1.7
cycler==0.11.0
Cython==0.29.28
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
dm-tree==0.1.6
docopt==0.6.2
entrypoints==0.3
fastavro==1.4.10
ffmpeg==1.4
Flask==2.0.2
flatbuffers==2.0
fonttools==4.28.5
gast==0.4.0
gin-config==0.5.0
google-api-core==2.5.0
google-api-python-client==2.41.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-speech==2.12.0
google-pasta==0.2.0
google-trans-new==1.1.9
googleapis-common-protos==1.55.0
googletrans==3.1.0a0
googletranslate==2.1.2
grpcio==1.44.0
grpcio-status==1.44.0
h11==0.9.0
h2==3.2.0
h5py==3.6.0
hdfs==2.6.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httplib2==0.19.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imbalanced-learn==0.9.0
imblearn==0.0
importlib-metadata==4.10.0
importlib-resources==5.4.0
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.3.3
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaggle==1.5.12
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
libcst==0.4.1
librosa==0.8.1
llvmlite==0.37.0
lvis==0.5.3
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.6.7
notebook==6.4.6
numba==0.54.1
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.1.1
object-detection @ file:///C:/Users/jocai/Downloads/tf_model_garden/research
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opt-einsum==3.3.0
orjson==3.6.7
packaging==21.3
pafy==0.5.5
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.0
pooch==1.5.2
portalocker==2.4.0
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.24
proto-plus==1.20.3
protobuf==3.19.2
psutil==5.9.0
py-cpuinfo==8.0.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
pydot==1.4.2
pydub==0.25.1
Pygments==2.11.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2021.3
pywin32==303
pywinpty==1.1.6
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.0
resampy==0.2.2
rfc3986==1.5.0
rsa==4.8
sacrebleu==2.0.0
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
sentencepiece==0.1.96
seqeval==1.2.2
six==1.16.0
sniffio==1.2.0
spyder-kernels==2.2.1
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-addons==0.16.1
tensorflow-datasets==4.5.2
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-hub==0.12.0
tensorflow-io==0.24.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-metadata==1.7.0
tensorflow-model-optimization==0.7.2
tensorflow-text==2.8.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
tf-models-official==2.8.0
tf-slim==1.1.0
thop==0.0.31.post2005241907
threadpoolctl==3.0.0
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typeguard==2.13.3
typing-inspect==0.7.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
youtube-dl==2021.12.17
zipp==3.7.0
